+++ CHANGELOG.md
@@ -9,7 +9,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+* Added `compas_rhino.artists.BoxArtist.draw_collection`.
+
 ### Changed
+
+* Changed parameters of `compas_rhino.artists.BoxArtist.draw`.
 
 ### Removed